requirements.txt
streamlit
requests
python-dotenv
pandas
st-aggrid
chardet
charset_normalizer
